+++ CHANGELOG.md
@@ -60,6 +60,7 @@ ___Note:__ Yet to be released changes appear here._
 * `FIX`: detect agentic ad-hoc sub-process by AI Agent job worker type ([camunda/bpmnlint-plugin-camunda-compat#260](https://github.com/camunda/bpmnlint-plugin-camunda-compat/pull/260))
 * `FIX`: label invalid `fromAi()` input as `Input from agent` instead of `Autofill` ([#6145](https://github.com/camunda/camunda-modeler/issues/6145))
 * `FIX`: display a uniform message for credentials when no connection is configured or established ([#6212](https://github.com/camunda/camunda-modeler/pull/6212))
+* `FIX`: display a message pointing to the gRPC connection when credentials are unavailable ([#6217](https://github.com/camunda/camunda-modeler/issues/6217))
 
 ### Forms